Work to support PR #734. Renamed code-export-utils to side-utils, which will be a public package. Placed the code-export specific bits behind `codeExport` so it can be destructured.
