+++ packages/code-export-python-pytest/__test__/src/index.spec.js
@@ -28,7 +28,7 @@ function readFile(filename) {
         '..',
         '..',
         '..',
-        'code-export-utils',
+        'side-utils',
         '__test__',
         'test-files',
         filename
