+++ packages/side-utils/__test__/src/code-export/preprocessor.spec.js
@@ -19,7 +19,7 @@ import {
   defaultPreprocessor,
   scriptPreprocessor,
   keysPreprocessor,
-} from '../../src/preprocessor'
+} from '../../../src/code-export/preprocessor'
 
 describe('preprocessors', () => {
   const variableLookup = varName => {
