+++ packages/side-utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "code-export-utils",
+  "name": "@seleniumhq/side-utils",
   "version": "3.10.0",
   "description": "Utils for Selenium IDE code export",
   "repository": "https://github.com/SeleniumHQ/selenium-ide",
@@ -9,7 +9,7 @@
     "export"
   ],
   "license": "Apache-2.0",
-  "private": true,
+  "private": false,
   "scripts": {
     "build": "babel -d dist src --ignore=__test__",
     "watch": "yarn build -w"
